for #2567, move SelectStatement.setIndexForItems to ShardingSelectOptimizedStatement
